separate type.h to directory types
