+++ src/index/concurrent_table.h
@@ -24,7 +24,9 @@
 #include <string_view>
 
 #include "concurrent_point_index_base.h"
-#include "types.h"
+#include "types/data_item.hpp"
+#include "types/definitions.h"
+#include "types/snapshot.hpp"
 
 namespace LineairDB {
 namespace Index {
